+++ src/pages/IndividualFeed.jsx
@@ -1,18 +1,13 @@
 import React, { useEffect, useState } from "react";
 import backgroundImage from "../assets/images/IndividualFeed-BackgroundImage.png";
 import logo from "../assets/images/logo.png";
-import facebook from "../assets/images/Facebook.png";
-import kakao from "../assets/images/Kakaotalk.png";
 import message from "../assets/images/Messages.png";
-import link from "../assets/images/Link.png";
 import FeedCard from "../components/FeedCard";
 import AddQuestion from "../components/AddQuestion";
 import {
   Wrapper,
   Logo,
   Profile,
-  Icons,
-  Icon,
   BodyWrapper,
   ProfilePlaceholder,
 } from "../styles/individualFeedStyle";
@@ -21,7 +16,7 @@ import { useIndividualQuestions } from "../hooks/useIndividualQuestions";
 import FeedCardPlaceholder from "../components/FeedCardPlaceholder";
 import { RotatingAnimation } from "../styles/rotatingAnimation";
 import { useScroll } from "../hooks/useScroll";
-import { useKakaoShare } from "../hooks/useKakaoShare";
+import IconBox from "../components/IconBox";
 
 const IndividualFeed = () => {
   const [offset, setOffset] = useState(0);
@@ -33,15 +28,6 @@ const IndividualFeed = () => {
   });
   const { moreData } = useScroll({ setOffset, questionInfo, LIMIT });
   const [loading, setLoading] = useState(true);
-  const { shareKakao } = useKakaoShare();
-  const facebookShareUrl = `https://www.facebook.com/sharer/sharer.php?u=${encodeURIComponent(
-    window.location.href
-  )}`;
-
-  // clipboard에 현재 url 복사
-  const copyUrl = () => {
-    navigator.clipboard.writeText(window.location.href);
-  };
 
   // 스켈리톤 ui를 위한 상태 변경
   useEffect(() => {
@@ -64,20 +50,8 @@ const IndividualFeed = () => {
       )}
       <span className="profileName">{userInfo.name}</span>
 
-      <Icons>
-        <Icon colorType="link" onClick={copyUrl}>
-          <img src={link} alt="링크" />
-        </Icon>
-        <Icon
-          onClick={() => shareKakao(window.location.href)}
-          colorType="kakao"
-        >
-          <img src={kakao} alt="카카오" />
-        </Icon>
-        <Icon href={facebookShareUrl} colorType="facebook">
-          <img src={facebook} alt="페이스북" />
-        </Icon>
-      </Icons>
+      {/* 아이콘 컴포넌트 */}
+      <IconBox />
 
       {/* 질문을 보여주는 부분 */}
       <BodyWrapper>
